refactor oauth into controller
